This changes the behavior of pull --overwrite mentioned in Issue #16.
Previously the code would remove any file that was soon to be
downloaded with --pull, if --overwrite was set. At the time, this
was intentional to save disk space and not double the usage temporarily.
 However, as this issue points out, an upstream API bug will prevent
a download from replacing these. The previous behavior was not wrong,
but is a bit less safe. Users wishing to avoid disk space usage can
manually remove files beforehand, or eventaully a --remove option can
be added.