Pull only a single file or a whole folder.
This is set using -l or --limit flag.
If a single file is given, it should be the full path (from the repo)
otherwise it won't match.
If a directory is given, match all files and subdirectories. If a
subdirectory is given, it should have all parents from the repo root.

Split the generation of a download queue in `pull` function into a
separate function for readability.